Remove certification process
A lot of people consistently misunderstood the certification process.
The goal was not to have a list of "approved" software, or software that is TT.
The goal was to help people looking for codebases that *they* will be able to
understand within a reasonable amount of time, by suggesting ones that someone
checked already.

The reason I was against self-certification is as follows:
1. The writer of a piece of software, at the time of writing, cannot be
   objective when measuring how long it would take to understand. This is
   because they already have the needed information in their head, making
   their certification worthless for the given purpose.
2. When you put the TT blurb in your README, you are *already* self-certifying.
   You are already claiming that your software is TT, and consequently no
   further listing or "certification" is needed - it is there and plain to see.
   The only further use would be to measure how easily someone fresh to the
   codebase could grok it - something that (as per 1) the author physically
   cannot do.

The unfortunate consequence of the misunderstandings has been forks that
loudly proclaim that in *their* version, you *can* self-certify and stick a
list of all the software you've made on your own page (or someone else's).
This is not only missing the point, but actively detracts from much more
important content elsewhere.

I would rather no have a silly "certified" list at all than waste the time of
people that may not have a large amount of it to spare to begin with.

And this is why it's being removed.
If I come up with better wording, I may bring it back.